HOTT-4457: Reflect new yielding find_routes api in rails
Rails 7.1.x moved to a yielding api for the internal find_routes method.

This broke our routing filter implementation by forcing us to reset the
path after controller dispatch had happened which was triggering
the forgery protection callback to be invoked with the wrong path.

We can fix this out of sequence issue, by resetting the path info
_before_ we transition to do controller dispatch.